Added support for supplementary characters.
- Introduced PropertyCharacterUnicode
- Added test cases associated to the new class
- Updated existing code to use the new class

+++ proactive-p2p-structured/proactive-p2p-structured-can/src/test/java/org/objectweb/proactive/extensions/p2p/structured/messages/can/AnycastLookupRequestTest.java
@@ -84,7 +84,7 @@ public class AnycastLookupRequestTest extends JunitByClassCanNetworkDeployer {
     public void testAnycastRequestWithResponse() {
         StringElement elt =
                 new StringElement(
-                        Character.toString((char) ((int) P2PStructuredProperties.CAN_UPPER_BOUND.getValue() - 1)));
+                        Character.toString((char) (P2PStructuredProperties.CAN_UPPER_BOUND.getValue() - 1)));
 
         GetZonesValidatingConstraintsResponse response =
                 (GetZonesValidatingConstraintsResponse) PAFuture.getFutureValue(this.proxy.send(
